requirements.txt
Pillow
pandas
scikit-learn
torch
torchvision
tqdm